requirements.txt
amqp==5.1.0
anyio==3.5.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.0
asttokens==2.0.5
async-timeout==4.0.2
atomicwrites==1.4.0
attrs==21.4.0
backcall==0.2.0
beautifulsoup4==4.11.1
billiard==3.6.4.0
bleach==5.0.0
blis==0.7.7
boto3==1.21.27
botocore==1.24.27
cachetools==5.0.0
catalogue==2.0.7
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cloudpickle==2.0.0
colorama==0.4.4
cycler==0.11.0
cymem==2.0.6
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
entrypoints==0.4
executing==0.8.3
fastapi==0.75.0
fastjsonschema==2.15.3
fonttools==4.33.3
gevent==21.12.0
google-api-core==2.5.0
google-auth==2.6.0
google-cloud-speech==2.12.0
googleapis-common-protos==1.54.0
greenlet==1.1.2
grpcio==1.43.0
grpcio-status==1.43.0
h11==0.13.0
hashids==1.3.1
httptools==0.3.0
huey==2.4.3
idna==3.3
iniconfig==1.1.1
ipykernel==6.13.0
ipython==8.3.0
ipython-genutils==0.2.0
ipywidgets==7.7.0
jedi==0.18.1
Jinja2==3.1.1
jiwer==2.3.0
jmespath==1.0.0
joblib==1.1.0
jsonschema==4.5.1
jupyter-client==7.3.0
jupyter-core==4.10.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.0
kiwisolver==1.4.2
kombu==5.2.4
langcodes==3.3.0
libcst==0.4.1
llvmlite==0.38.0
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
mistune==0.8.4
murmurhash==1.0.7
mypy-extensions==0.4.3
nbclient==0.6.2
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
nl-core-news-lg @ https://github.com/explosion/spacy-models/releases/download/nl_core_news_lg-3.2.0/nl_core_news_lg-3.2.0-py3-none-any.whl
notebook==6.4.11
numba==0.55.1
numpy==1.21.6
packaging==21.3
pandas==1.4.2
pandocfilters==1.5.0
parso==0.8.3
pathy==0.6.1
patsy==0.5.2
pickleshare==0.7.5
pika==1.2.0
Pillow==9.1.0
pip-licenses==3.5.3
pluggy==1.0.0
praat-parselmouth==0.4.1
preshed==3.0.6
prometheus-client==0.14.1
prompt-toolkit==3.0.29
proto-plus==1.20.0
protobuf==3.19.4
psutil==5.9.0
psycopg2==2.9.3
psycopg2-binary==2.9.3
PTable==0.9.2
pure-eval==0.2.2
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.7.4
Pygments==2.12.0
PyJWT==2.3.0
pyparsing==3.0.7
pyphen==0.12.0
pyrsistent==0.18.1
pytest==7.0.1
python-dateutil==2.8.2
python-dotenv==0.19.2
python-Levenshtein==0.12.2
python-multipart==0.0.5
pytz==2022.1
pywin32==304
pywinpty==2.0.5
PyYAML==6.0
pyzmq==22.3.0
queries==2.1.1
redis==4.2.2
requests==2.27.1
rsa==4.8
s3transfer==0.5.2
scikit-learn==1.0.2
scikit-plot==0.3.7
scipy==1.8.0
Send2Trash==1.8.0
shap==0.40.0
six==1.16.0
sklearn==0.0
slicer==0.0.7
smart-open==5.2.1
sniffio==1.2.0
soupsieve==2.3.2.post1
spacy==3.2.4
spacy-legacy==3.0.9
spacy-loggers==1.0.2
spacy-nightly==3.0.0rc5
srsly==2.4.3
stack-data==0.2.0
starlette==0.17.1
statsmodels==0.13.2
terminado==0.13.3
thinc==8.0.15
threadpoolctl==3.1.0
tinycss2==1.1.1
tomli==2.0.1
tornado==6.1
tqdm==4.64.0
traitlets==5.1.1
typer==0.3.2
typing-inspect==0.7.1
typing_extensions==4.1.1
urllib3==1.26.8
uvicorn==0.17.5
vine==5.0.0
vosk==0.3.32
wasabi==0.9.1
watchdog==2.1.7
watchdog-gevent==0.1.1
watchgod==0.7
wcwidth==0.2.5
webencodings==0.5.1
websockets==10.2
widgetsnbextension==3.6.0
wrapt==1.14.0
zope.event==4.5.0
zope.interface==5.4.0
